Unsecure cucumber token (#18)
* unsecure_cucumber_token
Due to a limitation on JUnit5 as runner, it cannot receive the token to publish the cucumber reports. See: https://github.com/cucumber/cucumber-jvm/issues/2409.
To avoid dealing with the issue, I'm moving the token to junit-platform.properties, not secure, but it works.

* unsecure_cucumber_token
Due to a limitation on JUnit5 as runner, it cannot receive the token to publish the cucumber reports. See: https://github.com/cucumber/cucumber-jvm/issues/2409.
To avoid dealing with the issue, I'm moving the token to junit-platform.properties, not secure, but it works.